refactor: extract theme-detector into useThemeDetector hook
Move the theme-application logic (data-theme sync, matchMedia listener
management for system mode) from SettingsContext's inline useEffect
into a dedicated useThemeDetector hook under src/hooks/.

The hook accepts a ThemeMode ('light' | 'dark' | 'system') and handles:
- Resolving the effective theme (querying prefers-color-scheme in system mode)
- Writing data-theme on document.documentElement
- Subscribing to OS preference changes when mode is 'system'
- Cleaning up the listener on mode switch or unmount

Behavior is identical — the change is a no-op refactor.

Closes #950
